pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "imperialist_agents"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "httpx>=0.27",
  "typer>=0.12",
  "pyyaml>=6.0",
  "rich>=13.7",
]

[project.optional-dependencies]
openai = ["openai>=1.47"]
llama = ["llama-index>=0.11"]

[project.scripts]
msagent = "manus_style_agent.cli:app"

[tool.setuptools.packages.find]
where = ["src"]